+++ game/classes/character.py
@@ -1,8 +1,5 @@
-from events import DialogMessage, StatusMessage
-from item import *
-import random
-import math
-
+import random, math
+from game.events.events import *
 
 class Character:
     _maxhp = 100
@@ -382,310 +379,3 @@ class Character:
         if not have_items:
             itemlist += "Your inventory is empty."
         return itemlist
-
-
-class Mage(Character):
-    _hp = 85
-    _mp = 10
-    _attack = 8
-
-    def __init__(self):
-        super().__init__()
-        self._cls = 'Mage'
-        self._es = self.get_es()
-        self.add_skill(Fireball(self))
-        self._passives['Energy Shield'] = "This passive allows Mage to absorb some of incoming damage.\n" \
-                                          "ES scales with MP and lvl"
-
-#   ES getter and setter #
-
-    def get_es(self):
-        """
-        Returns character's es
-        """
-        return self._mp * (self.get_lvl() / 2)
-
-    def _set_es(self, es):
-        """
-        Takes new value for character's es and sets it
-        Not really intended to use
-        """
-        self._es = es
-
-#   Class specific methods modifications #
-
-    def lvlup(self):
-        """
-        Gives character a lvlup bonus
-        """
-        super().lvlup()
-        self._es = self.get_es()
-
-    def take_damage_from(self, damage, other):
-        """
-        Takes other character's attack and reduces self es or/and hp by it
-        Prints message about that attack
-        """
-        if self._es > 0 and damage < self._es:
-            self._es -= damage
-            return DialogMessage('attack_es_CAT', other, damage, self).get_message() + "\n"
-        elif 0 < self._es <= damage:
-            if self._es == damage:
-                self._es = 0
-                return DialogMessage('broke_es_C', self).get_message() + "\n"
-            else:
-                leftoverdmg = damage - self._es
-                self._hp -= leftoverdmg * self.get_attack_modifier()
-                self._es = 0
-                return DialogMessage('broke_es_dmg_hp_CAT', other, leftoverdmg * self.get_attack_modifier(), self).get_message() + "\n"
-        else:
-            return super().take_damage_from(damage, other)
-
-    def get_stats(self):
-        """
-        Prints character's stats
-        """
-        stats = super().get_stats()
-        stats['ES'] = self._es
-        return stats
-
-
-class Warrior(Character):
-    _hp = 100
-    hp_mult = 1.1
-
-    def __init__(self):
-        super().__init__()
-        self._cls = 'Warrior'
-        self._maxhp = self._hp = self.get_maxhp()
-        self._passives['Warrior Blood'] = "This passive adds Warrior additional defence for every missing HP.\n"
-        self._passives['Great Health'] = "This passive adds additional defence for Warrior.\n"
-
-#   HP modifier, but with passive #
-
-    def get_hp_modifier(self):
-        """
-        Returns character's attack modifier
-        """
-        return self._hp_item_bonus
-
-    def get_maxhp(self):
-        """
-        Returns character's maxhp
-        """
-        return (self._maxhp + self.get_hp_modifier()) * self.hp_mult
-
-#   Armour passive getter #
-
-    def get_passive_defence_bonus(self):
-        """
-        Returns warrior's passive defence bonus
-        """
-        return math.log10(self.get_maxhp() - self.get_current_hp() + 1) / 10
-
-#   Class specific methods modifications #
-
-    def get_defence_modifier(self):
-        """
-        Returns character's defence modifier
-        """
-        return 1 - (math.log10(self.get_armour() + 1) / 3) - self.get_passive_defence_bonus()
-
-    def get_stats(self):
-        """
-        Returns character's stats in a dictionary
-        """
-        stats = super().get_stats()
-        stats['DEF_BONUS'] = round(self.get_passive_defence_bonus(), 2)
-        stats['HP_BONUS'] = self.hp_mult
-        return stats
-
-
-class Rogue(Character):
-    _crit_chance = 0.2
-    _evade_chance = 0.2
-    _attack = 15
-
-    def __init__(self):
-        super().__init__()
-        self._cls = 'Rogue'
-        self._evade_chance = self._evade_chance + 0.001 * self.get_mp()
-        self._passives['Evasion'] = "This passive allows Rogue to evade some of incoming damage.\n" \
-                                    "EV scales with MP"
-        self._passives['Critical Strike'] = "This passive allows Rogue to double the damage some of his attacks.\n" \
-                                            "Crit chance scales with MP"
-
-#   Evasion getters and setters #
-
-    def get_evasion(self):
-        """
-        Returns character's evade chance
-        """
-        return self._evade_chance + 0.001 * self.get_mp()
-
-    def set_evasion(self, ev):
-        """
-        Sets character's evade chance
-        """
-        self._evade_chance = ev
-
-    def get_dodge(self):
-        """
-        Returns evasion proc True or False
-        """
-        return random.random() < self.get_evasion()
-
-#   Crit getters and setters #
-
-    def get_crit_chance(self):
-        """
-        Returns character's evade chance
-        """
-        return self._crit_chance + 0.001 * self.get_mp()
-
-    def set_crit_chance(self, chance):
-        """
-        Takes new value for character's evade chance and sets it
-        """
-        self._crit_chance = chance
-
-    def get_crit(self):
-        """
-        Returns crit proc True or False
-        """
-        return random.random() < self.get_crit_chance()
-
-#   Class specific methods modifications #
-
-    def take_damage_from(self, damage, other):
-        """
-        Takes damage from other character
-        Prints message about that attack
-        """
-        if self.get_dodge():
-            return DialogMessage('evaded_CA', self, damage).get_message() + "\n"
-        else:
-            return super().take_damage_from(damage, other)
-
-    def attack(self, other):
-        """
-        Reduces other character's hp by self's attack
-        """
-        if self.get_crit():
-            attack = self.get_attack() * self.get_attack_modifier() * 2
-            return DialogMessage('crit', self, attack).get_message() + "\n" + other.take_damage_from(attack, self)
-        else:
-            attack = self.get_attack() * self.get_attack_modifier()
-            return other.take_damage_from(attack, self)
-
-    def add_item(self, item):
-        """
-        Adds an item to character inventory
-        """
-        if item is None:
-            pass
-        else:
-            if item.get_type() == 'Weapon':
-                item.set_name('Dagger')
-            super().add_item(item)
-
-    def get_stats(self):
-        """
-        Returns character's stats in a dictionary
-        """
-        stats = super().get_stats()
-        stats['EV_CHANCE'] = self.get_evasion()
-        stats['CRIT_CHANCE'] = self.get_crit_chance()
-        return stats
-
-
-class Monster(Character):
-    def __init__(self, lvl_mult=1):
-        super().__init__()
-        self._cls = 'Monster'
-        self._lvl_mult = lvl_mult / math.sqrt(lvl_mult * 2)
-        self._maxhp = (random.randint(21, 40) * self._lvl_mult)
-        self._hp = self.get_maxhp()
-        self._mp = (random.randint(1, 1) * self._lvl_mult)
-        self._attack = (random.randint(1, 15) * self._lvl_mult)
-
-
-class GreaterMonster(Monster):
-    def __init__(self, lvl_mult=1):
-        lvl_mult *= 2 * math.log10(lvl_mult)
-        super().__init__(lvl_mult)
-        self._cls = 'Greater Monster'
-        self.add_item(RareItem(int(lvl_mult)))
-        self.add_item(RareItem(int(lvl_mult)))
-        self.add_item(RareItem(int(lvl_mult)))
-        self._hp = self.get_maxhp()
-        self._mp = self.get_mp()
-
-
-class Skill:
-    def __init__(self, character):
-        self._owner = character
-        self._cooldown = 0
-        self._current_cd = 0
-
-    def get_owner(self):
-        """
-        Returns skill's owner
-        """
-        return self._owner
-
-    def get_name(self):
-        """
-        Returns skill's name
-        """
-        return self.__class__.__name__
-
-    def get_cooldown_timer(self):
-        """
-        Returns spells's cd timer
-        """
-        return self._cooldown
-
-    def set_cooldown_timer(self, cd):
-        """
-        Sets spells's cd timer
-        """
-        self._cooldown = cd
-
-    def get_current_cd(self):
-        """
-        Returns spell's current cd
-        """
-        return self._current_cd
-
-    def set_current_cd(self, cd):
-        """
-        Set new amount for current cd
-        More then max cd = max cd
-        Less then 0 = 0
-        """
-        if cd < 0:
-            self._current_cd = 0
-        elif cd <= self._cooldown:
-            self._current_cd = cd
-        else:
-            self._current_cd = self._cooldown
-
-    def is_available(self):
-        """
-        Returns if spell is ready to use
-        """
-        return self._current_cd == 0
-
-
-class Fireball(Skill):
-    def __init__(self, character):
-        super().__init__(character)
-        self._cooldown = 5
-        self._current_cd = 0
-
-    def get_damage(self):
-        """
-        Returns spell's damage
-        """
-        return self._owner.get_mp() * 1.5
